feat: add laravel permissions

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('cpf_cnpj');
-            $table->integer('phone_number');
+            $table->unsignedBigInteger('cpf_cnpj');
+            $table->unsignedBigInteger('phone_number');
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
